requirements.txt
streamlit~=1.43.1
streamlit-code-editor